switch to pgx driver, refactor for cli
