+++ cmd/demo/multiflag.go
@@ -1,0 +1,14 @@
+package main
+
+import "strings"
+
+type multiFlag []string
+
+func (m *multiFlag) String() string {
+	return strings.Join(*m, ",")
+}
+
+func (m *multiFlag) Set(value string) error {
+	*m = append(*m, value)
+	return nil
+}
